+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ Added
 * Add support for more date format codes (days, weeks, months and years) in the ``file`` setting
 * Add new ``round_entries`` setting to customize entries duration rounding
 * Add installation instructions for Nix and Debian systems
+* Add error message when trying to create entries with negative duration (#113)
 
 Changed
 -------